back/requirements.txt
flask
python-dotenv
flask_resful